More de-coupling, using _renderCanvas mixins, pixi.js inherits pixi-modern
