+++ bundles/pixi.js/package.json
@@ -36,40 +36,10 @@
     "@pixi/canvas-prepare": "^5.0.0-alpha",
     "@pixi/canvas-renderer": "^5.0.0-alpha",
     "@pixi/canvas-sprite": "^5.0.0-alpha",
-    "@pixi/core": "^5.0.0-alpha",
-    "@pixi/accessibility": "^5.0.0-alpha",
-    "@pixi/constants": "^5.0.0-alpha",
-    "@pixi/display": "^5.0.0-alpha",
-    "@pixi/extract": "^5.0.0-alpha",
-    "@pixi/filter-alpha": "^5.0.0-alpha",
-    "@pixi/filter-blur": "^5.0.0-alpha",
-    "@pixi/filter-color-matrix": "^5.0.0-alpha",
-    "@pixi/filter-displacement": "^5.0.0-alpha",
-    "@pixi/filter-fxaa": "^5.0.0-alpha",
-    "@pixi/filter-noise": "^5.0.0-alpha",
-    "@pixi/graphics": "^5.0.0-alpha",
-    "@pixi/interaction": "^5.0.0-alpha",
-    "@pixi/loaders": "^5.0.0-alpha",
-    "@pixi/math": "^5.0.0-alpha",
-    "@pixi/mesh": "^5.0.0-alpha",
-    "@pixi/mixin-cache-as-bitmap": "^5.0.0-alpha",
-    "@pixi/mixin-get-child-by-name": "^5.0.0-alpha",
-    "@pixi/mixin-get-global-position": "^5.0.0-alpha",
-    "@pixi/mixin-app-loader": "^5.0.0-alpha",
-    "@pixi/particles": "^5.0.0-alpha",
-    "@pixi/polyfill": "^5.0.0-alpha",
-    "@pixi/prepare": "^5.0.0-alpha",
-    "@pixi/settings": "^5.0.0-alpha",
-    "@pixi/sprite": "^5.0.0-alpha",
-    "@pixi/spritesheet": "^5.0.0-alpha",
-    "@pixi/sprite-animated": "^5.0.0-alpha",
-    "@pixi/sprite-tiling": "^5.0.0-alpha",
-    "@pixi/text": "^5.0.0-alpha",
-    "@pixi/text-bitmap": "^5.0.0-alpha",
-    "@pixi/ticker": "^5.0.0-alpha",
-    "@pixi/utils": "^5.0.0-alpha",
+    "@pixi/canvas-sprite-tiling": "^5.0.0-alpha",
     "@internal/builder": "^5.0.0-alpha",
     "@internal/tester": "^5.0.0-alpha",
+    "pixi-modern": "^5.0.0-alpha",
     "rollup": "^0.50.0"
   }
 }
