--- bundles/pixi.js/src/filters.js
@@ -1,26 +1,0 @@
-/**
- * This namespace contains WebGL-only display filters that can be applied
- * to DisplayObjects using the {@link PIXI.DisplayObject#filters filters} property.
- * @example
- * // Create a new application
- * const app = new PIXI.Application();
- *
- * // Draw a green rectangle
- * const rect = new PIXI.Graphics()
- *     .beginFill(0x00ff00)
- *     .drawRect(40, 40, 200, 200);
- *
- * // Add a blur filter
- * rect.filters = [new PIXI.filters.BlurFilter()];
- *
- * // Display rectangle
- * app.stage.addChild(rect);
- * document.body.appendChild(app.view);
- * @namespace PIXI.filters
- */
-export * from '@pixi/filter-fxaa';
-export * from '@pixi/filter-noise';
-export * from '@pixi/filter-displacement';
-export * from '@pixi/filter-blur';
-export * from '@pixi/filter-color-matrix';
-export * from '@pixi/filter-alpha';
