+++ bundles/pixi.js/src/index.js
@@ -1,77 +1,33 @@
-// import polyfills. Done as an export to make sure polyfills are imported first
 import '@pixi/polyfill';
 
-// export core
-export * from '@pixi/core';
-export * from '@pixi/app';
-export * from '@pixi/sprite';
-export * from '@pixi/spritesheet';
-export * from '@pixi/text';
-export * from '@pixi/text-bitmap';
-export * from '@pixi/graphics';
-export * from '@pixi/sprite-animated';
-export * from '@pixi/sprite-tiling';
-export * from '@pixi/math';
-export * from '@pixi/constants';
-export * from '@pixi/display';
+export * from 'pixi-modern';
 export * from '@pixi/canvas-renderer';
 export * from '@pixi/canvas-graphics';
 export * from '@pixi/canvas-mesh';
 export * from '@pixi/canvas-sprite';
 
-// Setup renderer plugins
-import './plugins';
-import './plugins-canvas';
-
-import * as accessibility from '@pixi/accessibility';
-import * as interaction from '@pixi/interaction';
-import * as extract from '@pixi/extract';
-import * as loaders from '@pixi/loaders';
-import * as mesh from '@pixi/mesh';
-import * as prepare from '@pixi/prepare';
-// import * as particles from '@pixi/particles';
-import * as filters from './filters';
-import * as utils from '@pixi/utils';
-import * as ticker from '@pixi/ticker';
-import { settings } from '@pixi/settings';
+import { InteractionManager } from '@pixi/interaction';
+import { CanvasRenderer, autoDetectRenderer } from '@pixi/canvas-renderer';
+import { CanvasMeshRenderer } from '@pixi/canvas-mesh';
+import * as canvasExtract from '@pixi/canvas-extract';
+import { CanvasGraphicsRenderer } from '@pixi/canvas-graphics';
 import * as canvasPrepare from '@pixi/canvas-prepare';
-import * as extractExtract from '@pixi/canvas-extract';
+import { CanvasSpriteRenderer } from '@pixi/canvas-sprite';
+import { AccessibilityManager } from '@pixi/accessibility';
 
+CanvasRenderer.registerPlugin('accessibility', AccessibilityManager);
+CanvasRenderer.registerPlugin('extract', canvasExtract.CanvasExtract);
+CanvasRenderer.registerPlugin('graphics', CanvasGraphicsRenderer);
+CanvasRenderer.registerPlugin('interaction', InteractionManager);
+CanvasRenderer.registerPlugin('mesh', CanvasMeshRenderer);
+CanvasRenderer.registerPlugin('prepare', canvasPrepare.CanvasPrepare);
+CanvasRenderer.registerPlugin('sprite', CanvasSpriteRenderer);
+
+import { prepare, extract } from 'pixi-modern';
 Object.assign(prepare, canvasPrepare);
-Object.assign(extract, extractExtract);
+Object.assign(extract, canvasExtract);
 
-// Support canvas fallback for Application renderer
-import { autoDetectRenderer } from '@pixi/canvas-renderer';
 import { Application } from '@pixi/app';
 Application.prototype.createRenderer = autoDetectRenderer;
 
-// imported for side effect of extending the prototype only, contains no exports
-import '@pixi/mixin-cache-as-bitmap';
-import '@pixi/mixin-get-child-by-name';
-import '@pixi/mixin-get-global-position';
-import '@pixi/mixin-app-loader';
-
-// handle mixins now, after all code has been added
-utils.mixins.performMixins();
-
-/**
- * Alias for {@link PIXI.loaders.shared}.
- * @name loader
- * @memberof PIXI
- * @type {PIXI.loader.Loader}
- */
-export const loader = loaders.shared;
-
-export {
-    accessibility,
-    extract,
-    filters,
-    interaction,
-    loaders,
-    mesh,
-    // particles,
-    prepare,
-    utils,
-    ticker,
-    settings,
-};
+import '@pixi/canvas-sprite-tiling';
