+++ packages/canvas-extract/package.json
@@ -29,9 +29,9 @@
     "lib"
   ],
   "dependencies": {
-    "@pixi/canvas-renderer": "^5.0.0-alpha",
     "@pixi/core": "^5.0.0-alpha",
-    "@pixi/math": "^5.0.0-alpha"
+    "@pixi/math": "^5.0.0-alpha",
+    "@pixi/utils": "^5.0.0-alpha"
   },
   "devDependencies": {
     "@internal/builder": "^5.0.0-alpha",
