+++ packages/canvas-graphics/src/index.js
@@ -1,1 +1,3 @@
 export { default as CanvasGraphicsRenderer } from './CanvasGraphicsRenderer';
+
+import './Graphics';
