+++ packages/canvas-mesh/src/index.js
@@ -1,1 +1,4 @@
 export { default as CanvasMeshRenderer } from './CanvasMeshRenderer';
+
+import './NineSlicePlane';
+import './RawMesh';
