+++ packages/canvas-renderer/src/CanvasRenderer.js
@@ -1,6 +1,6 @@
 import { AbstractRenderer } from '@pixi/core';
+import { CanvasRenderTarget } from '@pixi/utils';
 import CanvasMaskManager from './utils/CanvasMaskManager';
-import CanvasRenderTarget from './utils/CanvasRenderTarget';
 import mapCanvasBlendModesToPixi from './utils/mapCanvasBlendModesToPixi';
 import { RENDERER_TYPE, SCALE_MODES, BLEND_MODES } from '@pixi/constants';
 import { settings } from '@pixi/settings';
