+++ packages/canvas-renderer/src/index.js
@@ -1,4 +1,3 @@
 export { default as CanvasRenderer } from './CanvasRenderer';
 export { default as canUseNewCanvasBlendModes } from './utils/canUseNewCanvasBlendModes';
-export { default as CanvasRenderTarget } from './utils/CanvasRenderTarget';
 export { autoDetectRenderer } from './autoDetectRenderer';
