+++ packages/canvas-sprite-tiling/README.md
@@ -1,0 +1,13 @@
+# @pixi/canvas-sprite-tiling
+
+## Installation
+
+```bash
+npm install @pixi/canvas-sprite-tiling
+```
+
+## Usage
+
+```js
+import '@pixi/canvas-sprite-tiling';
+```
