+++ packages/canvas-sprite/package.json
@@ -30,6 +30,7 @@
   ],
   "dependencies": {
     "@pixi/canvas-renderer": "^5.0.0-alpha",
+    "@pixi/sprite": "^5.0.0-alpha",
     "@pixi/constants": "^5.0.0-alpha",
     "@pixi/math": "^5.0.0-alpha",
     "@pixi/utils": "^5.0.0-alpha"
