+++ packages/canvas-sprite/src/index.js
@@ -1,2 +1,4 @@
 export { default as CanvasSpriteRenderer } from './CanvasSpriteRenderer';
 export { default as CanvasTinter } from './CanvasTinter';
+
+import './Sprite';
