+++ packages/extract/package.json
@@ -29,7 +29,7 @@
     "lib"
   ],
   "dependencies": {
-    "@pixi/canvas-renderer": "^5.0.0-alpha",
+    "@pixi/utils": "^5.0.0-alpha",
     "@pixi/core": "^5.0.0-alpha",
     "@pixi/math": "^5.0.0-alpha"
   },
