+++ packages/sprite-tiling/package.json
@@ -29,8 +29,6 @@
     "lib"
   ],
   "dependencies": {
-    "@pixi/canvas-renderer": "^5.0.0-alpha",
-    "@pixi/canvas-sprite": "^5.0.0-alpha",
     "@pixi/core": "^5.0.0-alpha",
     "@pixi/math": "^5.0.0-alpha",
     "@pixi/utils": "^5.0.0-alpha",
