+++ packages/utils/src/media/index.js
@@ -1,2 +1,3 @@
 export * from './caches';
 export * from './trimCanvas';
+export { default as CanvasRenderTarget } from './CanvasRenderTarget';
