+++ tools/tester/index.js
@@ -13,8 +13,6 @@ floss(options, (err) =>
 {
     if (err)
     {
-        // eslint-disable-next-line no-console
-        console.log(err);
         process.exit(1);
     }
     else
